Add files via upload
zip file for my new gaming mouse a4tech x7 x-710bk oscar gaming.
Now heres the problem, there is some zipfiles are here and there for windows but wine does not seem to integrate the properly with the drivers. Imma try some vms but if that dosent work re-engeneering it is. I know that it has its own c like scripting language/library so some ide should get through it or the ArduinoIDE can try it.